% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{customize}
\alias{customize}
\title{Basic Scatter Plot Editor}
\usage{
customize(
  pt.type,
  pt.col = "black",
  pt.size = 1,
  line.choice = F,
  line.type = 1,
  line.col = "black",
  line.method = "default",
  change.lab = F,
  x.newlab = .xname,
  y.newlab = .yname,
  title.choice = F,
  new.title = "",
  new.subtitle = ""
)
}
\arguments{
\item{pt.type}{Required parameter: Point type based on R Base point types (expected input range: integers 0-25).}

\item{pt.col}{Optional parameter: Character name of color for points based on R Base colors. [Default is "black"]}

\item{pt.size}{Optional parameter: Numeric for size of points [Default is 1]}

\item{line.choice}{Optional parameter: Boolean decision of including a line. Note that this must be called to make changes in other variables involving the lines in the plot. [Default is F or FALSE]}

\item{line.type}{Optional parameter: Numeric indicating the line type based on R Base values for line types. [Default is 1 (solid line)]}

\item{line.col}{Optional parameter: Character/ string indicating color change of line in plot based on R Base colors. [Default is "black"]}

\item{line.method}{Optional parameter: Character/ string indicating whether to include a line of best fit or a default line connecting all points (expected input is "default" or "linear"). [Default is "default"]}

\item{change.lab}{Optional parameter: Boolean decision of including changing axes labels. Note if x.newlab() and y.newlab() are unchanged, the default parameter applies. [Default is F or FALSE]}

\item{x.newlab}{Optional parameter: Character/ string of x axis title label. [Default is the previous name from original plot]}

\item{y.newlab}{Optional parameter: Character/ string of y axis title label. [Default is the previous name from original plot]}

\item{title.choice}{Optional parameter: Boolean decision of including a main title for plot. [Defaault is F or FALSE]}

\item{new.title}{Optional parameter: Character/ string of main title in plot. [Default is ""]}

\item{new.subtitle}{Optional parameter: Character/ string of subtitle in plot. [Default is ""]}
}
\description{
customize() allows users to change the features of a current plot without rebuilding the plot. The input data is intended to include basic editting features of the basic scatter plot.
}
\examples{
#####the scatterplot.basic() plot must be called before use
# of the customize() function
scatterplot.basic(dat = mtcars, xname = "mpg", yname = "wt")

#####the customize() function now can be used to add/
#change plot features
# *can change the point type to triangles if desired
# (value of 2 in R Base)
# customize(pt.type = 2)

# *can change the color and size of the points
# (note that background color for input 21-25 is "red" by default)
customize(pt.type = 2, pt.col = "red", pt.size = 1)

# *can include a line and titles
customize(pt.type = 2, pt.col = "red", pt.size = 1,
line.choice = T, line.type = 1, title.choice = T,
new.title = "Testing out the customize function",
new.subtitle = "It looks like it's working.",
line.method = "linear")

# *can change each input
customize(pt.type = 2, pt.col = "red", pt.size = 1,
line.choice = T, line.type = 1, title.choice = T,
new.title = "Testing out the customize function",
new.subtitle = "It looks like it's working.",
line.method = "linear", line.col = "blue",
change.lab = T, x.newlab = "x-axis",
y.newlab = "y-axis")

}
