% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot.basic}
\alias{scatterplot.basic}
\title{Create a Basic Scatter Plot}
\usage{
scatterplot.basic(dat, xname, yname)
}
\arguments{
\item{dat}{a data frame}

\item{xname}{a character variable}

\item{yname}{a character variable}
}
\description{
scatterplot.basic() creates a plot including a grid similar to the use of ggplot with the geom_point function. It is used to visualize data specifically in a basic scatter plot. Note that four global variables named "xdata", "ydata", "xname", and "yname" will be created. After use of this function, you may have to reset graphical parameters to default using the par() function in the Graphics R package.
}
\examples{
my.data <- data.frame(cbind(1:5,2:6,3:7))
scatterplot.basic(my.data,"X1","X3")

}
